python-lambda-simple/requirements.txt
attrs==19.3.0
aws-cdk.assets==1.32.2
aws-cdk.aws-cloudwatch==1.32.2
aws-cdk.aws-ec2==1.32.2
aws-cdk.aws-events==1.32.2
aws-cdk.aws-iam==1.32.2
aws-cdk.aws-kms==1.32.2
aws-cdk.aws-lambda==1.32.2
aws-cdk.aws-logs==1.32.2
aws-cdk.aws-s3==1.32.2
aws-cdk.aws-s3-assets==1.32.2
aws-cdk.aws-sqs==1.32.2
aws-cdk.aws-ssm==1.32.2
aws-cdk.core==1.32.1
aws-cdk.cx-api==1.32.1
aws-cdk.region-info==1.32.2
cattrs==1.0.0
constructs==2.0.1
jsii==1.1.0
publication==0.0.3
python-dateutil==2.8.1
-e git+git@github.com:ustrahlendorf/cdk-templates.git@b134998954307030b9cbfc6d6b250ac68fe279da#egg=python_lambda_simple&subdirectory=python-lambda-simple
six==1.14.0
typing-extensions==3.7.4.2
